Fixed  some comment issues
