+++ anomaly_detection/api/services.py
@@ -16,6 +16,9 @@ from flask import jsonify
 from flask import Blueprint
 # from flask import request
 from anomaly_detection import log
+from anomaly_detection.db import api
+from anomaly_detection.utils import uuid
+from anomaly_detection import context as ctx
 
 service = Blueprint("service", __name__)
 LOG = log.getLogger(__name__)
@@ -25,4 +28,5 @@ LOG = log.getLogger(__name__)
 @service.route("/v1beta", methods=['GET'])
 def get_version():
     LOG.debug("get anomaly detection version")
+
     return jsonify(name="Anomaly Detection", version="v1beta"), 200
